Check if anyone was home already and send notification to hello home
Checks list of people to see if someone was home so that we don't turn lights on again that others might have off on purpose (watching TV for instance). Checks to see if luminance sensor is `!= null` instead of using falsy value so that we can get a zero value (not sure if this happens in the real world but feels more correct). Sends notification to hello home so the user can see what happened and why.